telegram bot

+++ src/main/java/org/expense_bot/handler/categories/CategoryFinalActionHandler.java
@@ -8,18 +8,17 @@ import org.expense_bot.handler.RequestHandler;
 import org.expense_bot.handler.init.BackButtonHandler;
 import org.expense_bot.helper.KeyboardBuilder;
 import org.expense_bot.model.Category;
-import org.expense_bot.model.UserCategory;
 import org.expense_bot.model.Request;
+import org.expense_bot.model.UserCategory;
 import org.expense_bot.service.CategoryService;
 import org.expense_bot.service.UserCategoryService;
-import org.expense_bot.service.impl.TelegramService;
 import org.expense_bot.service.impl.SessionService;
+import org.expense_bot.service.impl.TelegramService;
 import org.expense_bot.util.SessionUtil;
 import org.springframework.stereotype.Component;
 import org.telegram.telegrambots.meta.api.objects.replykeyboard.ReplyKeyboard;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Component
@@ -45,22 +44,20 @@ public class CategoryFinalActionHandler extends RequestHandler {
 	final String param = getUpdateData(request);
 	final CategoryAction categoryAction = sessionService.getSession(userId).getCategoryAction();
 
-	Optional.ofNullable(categoryAction)
-	  .ifPresent(action -> {
-		switch (action) {
-		  case ADD_NEW_CATEGORY:
-			addCategory(userId, param);
-			break;
-		  case ADD_FROM_DEFAULT:
-			userCategoryService.add(userId, param);
-			sendListNotSelected(userId);
-			break;
-		  case DELETE_MY_CATEGORIES:
-			userCategoryService.delete(userId, param);
-			sendListNotDeleted(userId);
-			break;
-		}
-	  });
+	if(categoryAction != null) {
+	  switch (categoryAction) {
+		case ADD_NEW_CATEGORY:
+		  addCategory(userId, param);
+		  return;
+		case ADD_FROM_DEFAULT:
+		  userCategoryService.add(userId, param);
+		  sendListNotSelected(userId);
+		case DELETE_MY_CATEGORIES:
+		  userCategoryService.delete(userId, param);
+		  sendListNotDeleted(userId);
+		  break;
+	  }
+	}
 
 	sessionService.update(SessionUtil.getSession(userId, categoryAction));
   }
@@ -69,12 +66,14 @@ public class CategoryFinalActionHandler extends RequestHandler {
 	userCategoryService.add(chatId, param);
 	telegramService.sendMessage(chatId, Messages.CATEGORY_ADDED_TO_YOUR_LIST, keyboardBuilder.buildCategoryOptionsMenu());
 	sessionService.updateState(chatId, ConversationState.Categories.WAITING_CATEGORY_ACTION);
-	throw new RuntimeException(Messages.CATEGORY_ADDED_TO_YOUR_LIST);
   }
 
   private void sendListNotDeleted(Long chatId) {
 	final List<String> userCategories = getUserCategories(chatId);
-	checkAll(chatId, userCategories, Messages.ALL_CATEGORIES_DELETED);
+	final boolean checkAll = checkAll(chatId, userCategories, Messages.ALL_CATEGORIES_DELETED);
+	if(checkAll){
+	  return;
+	}
 	final ReplyKeyboard keyboard = keyboardBuilder.buildCustomCategoriesMenu(userCategories);
 	telegramService.sendMessage(chatId, Messages.ASK_TO_DELETE, keyboard);
   }
@@ -86,16 +85,19 @@ public class CategoryFinalActionHandler extends RequestHandler {
 	  .map(Category::getName)
 	  .collect(Collectors.toList());
 	defaultCategories.removeAll(userCategories);
-	checkAll(chatId, defaultCategories, Messages.ALL_CATEGORIES_ADDED);
+	final boolean checkAll = checkAll(chatId, defaultCategories, Messages.ALL_CATEGORIES_ADDED);
+	if(checkAll){
+	  return;
+	}
 	final ReplyKeyboard keyboard = keyboardBuilder.buildCustomCategoriesMenu(defaultCategories);
 	telegramService.sendMessage(chatId, Messages.CHOOSE_CATEGORY_FROM_LIST, keyboard);
   }
 
-  private void checkAll(Long chatId, List<String> categories, String message) {
+  private boolean checkAll(Long chatId, List<String> categories, String message) {
 	if(categories.isEmpty()) {
 	  telegramService.sendMessage(chatId, message, keyboardBuilder.buildBackButton());
-	  throw new RuntimeException(message);
 	}
+	return categories.isEmpty();
   }
 
   @Override
